Rework resolving references and namespace creation in modules
- Properties of namespace objects are ordered in lexical order
- The namespace object is created and imports are bound during linking phase
- Namespace imports are changed to read-only
- Reduced memory consumption

JerryScript-DCO-1.0-Signed-off-by: Zoltan Herczeg zherczeg.u-szeged@partner.samsung.com

+++ jerry-core/ecma/base/ecma-globals.h
@@ -1061,7 +1061,7 @@ typedef struct
       {
         uint16_t formal_params_number; /**< for arguments: formal parameters number */
 #if JERRY_MODULE_SYSTEM
-        uint8_t module_flags; /**< Module flags */
+        uint16_t module_flags; /**< Module flags */
 #endif /* JERRY_MODULE_SYSTEM */
 #if JERRY_ESNEXT
         uint16_t iterator_index; /**< for %Iterator%: [[%Iterator%NextIndex]] property */
